Use Neo4j API instead of Cypher queries to retrieve nodes

+++ src/main/java/name/matco/simcity/model/Ingredient.java
@@ -4,5 +4,6 @@ public class Ingredient {
 
 	public String id;
 	public String name;
-	public int time;
+	public Long time;
+
 }
